New reader.
This new reader should be both more lispy and a lot more efficient rust
code.  It uses a single iterator over the input (with single char
lookahead) and one buffer when needed to build strings or symbols.
Therefore there are a LOT less allocations.  There should still be room
for improvement here but this should be a LOT better.