Remove the Reference type and officially make 'Scopes' 'Namespaces'
This is still a WIP and this branch only compiles but will not run
correctly yet.  This should simplify a lot of stuff but doc strings can
only be set on a namespaced symbol now.  Not sure why you want to set a
doc string on a local symbol so this seems reasonable.